Moves Csharp-specific things to separate folder. Adds generator option to node api
