+++ lib/index.js
@@ -12,11 +12,17 @@ const getPropTypes = require('./utils/get-prop-types');
 const getPropTypesIdentifierName = require('./utils/get-prop-types-identifier-name');
 const parseAST = require('./utils/parse-ast');
 const stripPropTypesIdentifier = require('./utils/strip-prop-types-identifier');
-const stringify = require('./stringify');
+const generateCsharp = require('./stringify/lang/csharp');
 const transformPropTypes = require('./transforms/transform-prop-types');
 const validatePropNames = require('./utils/validate-prop-names');
 
-module.exports = function({ baseClass, indent, namespace, sourceCode }) {
+module.exports = function({
+  baseClass,
+  generator = generateCsharp,
+  indent,
+  namespace,
+  sourceCode
+}) {
   assert(sourceCode, 'No source code provided');
 
   const syntaxTree = parse(sourceCode, {
@@ -77,7 +83,7 @@ module.exports = function({ baseClass, indent, namespace, sourceCode }) {
 
   const definitions = createDefinitions(transformedPropTypes, componentName);
 
-  const code = stringify({
+  const code = generator({
     baseClass,
     componentName,
     definitions,
