+++ lib/stringify/lang/csharp/enum-properties.js
@@ -1,5 +1,5 @@
-const capitalize = require('./capitalize');
-const unknownToPascal = require('./unknown-to-pascal');
+const capitalize = require('../../capitalize');
+const unknownToPascal = require('../../unknown-to-pascal');
 
 // Returns the body of an enum definition as string
 const enumProperties = (name, { argument, isRequired }) => {
