+++ lib/stringify/lang/csharp/generator.js
@@ -1,8 +1,8 @@
-const capitalize = require('./capitalize');
+const capitalize = require('../../capitalize');
 
 const enumProperties = require('./enum-properties');
 const nestedArrayString = require('./nested-array-string');
-const parentPrefixedName = require('./parent-prefixed-name');
+const parentPrefixedName = require('../../parent-prefixed-name');
 
 const property = (name, { isRequired }, type) =>
   `${isRequired ? '[Required]\n' : ''}` +
