+++ lib/stringify/lang/csharp/index.js
@@ -1,5 +1,5 @@
-const addIndentation = require('./indent');
 const { generateClass, generateClassExtends } = require('./generator');
+const indentBraces = require('../../indent-braces');
 const imports = require('./imports');
 
 // Expects syntax tree consisting of only assignment expression nodes in the program.body node
@@ -26,5 +26,5 @@ module.exports = function({
 
   const fileContent = `${imports}\n${classesWithNamespace}`;
 
-  return addIndentation(fileContent, indent);
+  return indentBraces(fileContent, indent);
 };
