+++ lib/stringify/lang/csharp/nested-array-string.js
@@ -1,4 +1,4 @@
-const parentPrefixedName = require('./parent-prefixed-name');
+const parentPrefixedName = require('../../parent-prefixed-name');
 
 const nestedArrayString = (propType, parent, accum = '') => {
   if (propType.type === 'arrayOf') {
