+++ test/lib/nested-array-string.js
@@ -1,6 +1,6 @@
 const test = require('ava');
 
-const nestedArrayString = require('../../lib/stringify/nested-array-string');
+const nestedArrayString = require('../../lib/stringify/lang/csharp/nested-array-string');
 
 test('With external class', t => {
   const arrayString = nestedArrayString({
